misc: refactored files
